chore: 🤖 move toggletheme to app controller
