+++ ui/desktop/app/controllers/scopes/scope/projects/settings/index.js
@@ -5,7 +5,6 @@
 
 import Controller, { inject as controller } from '@ember/controller';
 import { inject as service } from '@ember/service';
-import { getOwner } from '@ember/application';
 import { action } from '@ember/object';
 
 const THEMES = [
@@ -24,11 +23,10 @@ const THEMES = [
 ];
 
 export default class ScopesScopeProjectsSettingsIndexController extends Controller {
-  @controller('scopes/scope/projects/settings/index') settings;
+  @controller('application') application;
   // Services
 
   @service session;
-  @service ipc;
 
   // Methods
 
@@ -40,31 +38,11 @@ export default class ScopesScopeProjectsSettingsIndexController extends Controll
   }
 
   /**
-   * Applies the specified color theme to the root ember element.
-   * @param {string} theme - "light", "dark", or nullish (system default) or an event object
+   * Calls the Application controller's toggleTheme method
+   * @param {string} theme - "light", "dark", or nullish (system default)
    */
   @action
-  toggleTheme(theme) {
-    if (typeof theme === 'object') {
-      theme = theme.target.value;
-    }
-    const rootElementSelector = getOwner(this).rootElement;
-    const rootEl = getOwner(this)
-      .lookup('service:-document')
-      .querySelector(rootElementSelector);
-    this.session.set('data.theme', theme);
-    switch (theme) {
-      case 'light':
-        rootEl.classList.add('rose-theme-light');
-        rootEl.classList.remove('rose-theme-dark');
-        break;
-      case 'dark':
-        rootEl.classList.add('rose-theme-dark');
-        rootEl.classList.remove('rose-theme-light');
-        break;
-      default:
-        rootEl.classList.remove('rose-theme-dark');
-        rootEl.classList.remove('rose-theme-light');
-    }
+  toggleTheme({ target: { value: theme } }) {
+    return this.application.toggleTheme(theme);
   }
 }
